imagewatcher-python/requirements.txt
Flask==0.10.1
gunicorn==19.4.5

# https://github.com/dpkp/kafka-python
kafka-python

# https://github.com/jmcarp/py-cfenv
cfenv
